피드 수정 requestDto의 peopleCount type 수정

+++ src/main/java/GDSC/FirstProject/dto/requsetDto/UpdateFeedRequestDto.java
@@ -15,7 +15,7 @@ public class UpdateFeedRequestDto {
 
     private String[] hashtags;
 
-    private Long peopleCount;
+    private Integer peopleCount;
 
     private String company;
